Storage: add volume type helper file
Moved some functions to a new file under
storage/types.
Those functions were either in the pvc types,
utils/volumes which handles different stuff
and under controller.
Adjusted imports accordingly and moved tests
