+++ pkg/storage/utils/volumes.go
@@ -198,12 +198,3 @@ func getNewestNonTerminatingPVC(pvcs []k8sv1.PersistentVolumeClaim) (*k8sv1.Pers
 func IsErrNoBackendPVC(err error) bool {
 	return errors.Is(err, ErrNoBackendPVC)
 }
-
-func IsConfigVolume(volume *v1.Volume) bool {
-	return volume.ConfigMap != nil || volume.Secret != nil ||
-		volume.ServiceAccount != nil || volume.DownwardAPI != nil
-}
-
-func IsStorageVolume(volume *v1.Volume) bool {
-	return volume.PersistentVolumeClaim != nil || volume.DataVolume != nil
-}
