+++ pkg/virt-controller/watch/vmi/lifecycle.go
@@ -185,7 +185,7 @@ func (c *Controller) sync(vmi *virtv1.VirtualMachineInstance, pod *k8sv1.Pod, da
 		}
 		pod = patchedPod
 
-		hotplugVolumes := controller.GetHotplugVolumes(vmi, pod)
+		hotplugVolumes := storagetypes.GetHotplugVolumes(vmi, pod)
 		hotplugAttachmentPods, err := controller.AttachmentPods(pod, c.podIndexer)
 		if err != nil {
 			return common.NewSyncError(fmt.Errorf("failed to get attachment pods: %v", err), controller.FailedHotplugSyncReason), pod
