+++ pkg/virt-controller/watch/vmi/vmi_test.go
@@ -2835,7 +2835,7 @@ var _ = Describe("VirtualMachineInstance watcher", func() {
 			}
 			virtlauncherPod := newPodForVirtualMachine(vmi, k8sv1.PodRunning)
 			virtlauncherPod.Spec.Volumes = virtlauncherVolumes
-			res := kvcontroller.GetHotplugVolumes(vmi, virtlauncherPod)
+			res := storagetypes.GetHotplugVolumes(vmi, virtlauncherPod)
 			Expect(res).To(HaveLen(len(expectedIndexes)))
 			for _, index := range expectedIndexes {
 				found := false
